+++ jasperreports/src/net/sf/jasperreports/engine/xml/JRPrintImageFactory.java
@@ -70,19 +70,19 @@ public class JRPrintImageFactory extends JRBaseFactory
 		Byte scaleImage = (Byte)JRXmlConstants.getScaleImageMap().get(atts.getValue(ATTRIBUTE_scaleImage));
 		if (scaleImage != null)
 		{
-			image.setScaleImage(scaleImage.byteValue());
+			image.setScaleImage(scaleImage);
 		}
 
 		Byte horizontalAlignment = (Byte)JRXmlConstants.getHorizontalAlignMap().get(atts.getValue(ATTRIBUTE_hAlign));
 		if (horizontalAlignment != null)
 		{
-			image.setHorizontalAlignment(horizontalAlignment.byteValue());
+			image.setHorizontalAlignment(horizontalAlignment);
 		}
 
 		Byte verticalAlignment = (Byte)JRXmlConstants.getVerticalAlignMap().get(atts.getValue(ATTRIBUTE_vAlign));
 		if (verticalAlignment != null)
 		{
-			image.setVerticalAlignment(verticalAlignment.byteValue());
+			image.setVerticalAlignment(verticalAlignment);
 		}
 
 		String isLazy = atts.getValue(ATTRIBUTE_isLazy);
